# perf(gatsby-plugin-sharp): change approach to concurrency for image processing