## Description

My tests for lazy images (#27603) showed that increasing sharp queue concurrency results in a better overall image transformation performance. Tested on two image-heavy sites.

First one (based on [this starter](https://github.com/baobabKoodaa/gatsby-starter-photo-book)):

Before this PR:
```
success Generating image thumbnails - 174.163s - 532/532 3.05/s
info Done building in 179.6475083 sec
```

After this PR:
```
success Generating image thumbnails - 311.836s - 532/532 1.71/s
info Done building in 317.3029042 sec
```

[Second site](https://github.com/kentcdodds/kentcdodds.com):

Before this PR:
```
Generating image thumbnails - 436.176s
```

After this PR:
```
Generating image thumbnails - 259.491s
```

Note that for this site the build has failed for me with unrelated SSR error - for both checks. Just don't have time to figure it out now - my fork is pretty outdated. But image processing was significantly faster this way.

I also see similar good effect on lazy images - images load about `30-40%` faster.

CC @pvdz @wardpeet